updating files

+++ RBL.GitHub.Scrapper.Domain/FileInformation.cs
@@ -6,16 +6,16 @@ namespace RBL.GitHub.Scrapper.Domain
 {
     public class FileInformations
     {
-        public FileInformations(int lines, long size)
+        public FileInformations(int lines, decimal size)
         {
             this._lines = lines;
             this._size = size;
         }
 
         private int _lines { get; set; }
-        private long _size { get; set; }
+        private decimal _size { get; set; }
 
         public int Lines { get { return _lines; } }
-        public long Size { get { return _size; } }
+        public decimal Size { get { return _size; } }
     }
 }
